requirements.txt
beaker-pyteal==0.1.5a0
cffi==1.15.1
msgpack==1.0.4
py-algorand-sdk==1.16.1
pycparser==2.21
pycryptodomex==3.15.0
PyNaCl==1.5.0
pyteal==0.17.0
semantic-version==2.10.0
